Refactor: Move tests to sql_composer_pg_test.pg. Created sql_composer_pg.py
